requirements.txt
openai>=1.0.0
rich>=13.0.0
pytest>=7.0.0
python-dotenv>=1.0.0 